chore: reorganize the whole folder structure to keep it more simple and only import raw functions to resolvers instead of services and controllers
