+++ src/graphql/resolvers/account.resolver.ts
@@ -1,5 +1,5 @@
-import accountController from "@/controllers/account.controller";
 import Account from "@/models/Account";
+import { transferAmount, accountByUserId } from "@/account";
 
 const accountResolver = {
   Query: {
@@ -7,12 +7,12 @@ const accountResolver = {
       return await Account.find(id);
     },
     async accountByUserId(_: any, { userId }: any) {
-      return await Account.findOne({ userId });
+      return await accountByUserId(userId);
     },
   },
   Mutation: {
     async transferAmount(_: any, args: any) {
-      return await accountController.transfer(args);
+      return await transferAmount(args);
     },
   },
 };
