+++ src/graphql/resolvers/index.ts
@@ -1,8 +1,8 @@
 import { mergeResolvers } from '@graphql-tools/merge'
 import { resolvers as scalarResolvers } from 'graphql-scalars'
-import accountResolver from './account-resolver'
+import accountResolver from './account.resolver'
 import transationResolver from './transaction.resolver'
-import userResolver from './user-resolver'
+import userResolver from './user.resolver'
 
 const resolvers = [
   userResolver,
