--- src/services/token.service.ts
@@ -1,37 +1,0 @@
-import { decodeToken, JWT_EXPIRES_IN } from "@/helpers/jwt";
-import Token from "@/models/Token";
-
-/**
- * Check if token exists
- * @param token {string} - Token
- * @returns {Promise<Boolean>} - True if exists, false otherwise
- */
-export const tokenExists = async (token: string): Promise<boolean> =>
-  Boolean(await Token.exists({ token }));
-
-/**
- *  Revoke a token
- * @param token
- * @returns
- */
-export const revokeToken = async (token: string) => {
-  if(!token) return false;
-  
-  const existingToken = await tokenExists(token);
-
-  if (existingToken) return true;
-
-  const decodedToken = decodeToken(token) || { exp: JWT_EXPIRES_IN };
-
-  const expiresAt = new Date(decodedToken.exp * 1000);
-
-  const blackListed = new Token({ token, expiresAt });
-
-  if (blackListed.isNew) {
-    await blackListed.save();
-
-    return true;
-  }
-
-  return false;
-};
